Move Coordinates and IVector to vector package
