+++ BaseWeapon/src/main/java/com/github/sef24sp4/baseweapon/ProjectileControlSystem.java
@@ -1,6 +1,6 @@
 package com.github.sef24sp4.baseweapon;
 
-import com.github.sef24sp4.common.data.Coordinates;
+import com.github.sef24sp4.common.vector.Coordinates;
 import com.github.sef24sp4.common.entities.IEntity;
 import com.github.sef24sp4.common.interfaces.IEntityManager;
 import com.github.sef24sp4.common.interfaces.IGameSettings;
