+++ Common/src/main/java/com/github/sef24sp4/common/entities/IEntity.java
@@ -1,6 +1,6 @@
 package com.github.sef24sp4.common.entities;
 
-import com.github.sef24sp4.common.data.Coordinates;
+import com.github.sef24sp4.common.vector.Coordinates;
 
 import java.util.ArrayList;
 import java.util.List;
