+++ Common/src/main/java/com/github/sef24sp4/common/gamecontrol/IMouseCoordinates.java
@@ -1,6 +1,6 @@
 package com.github.sef24sp4.common.gamecontrol;
 
-import com.github.sef24sp4.common.interfaces.IVector;
+import com.github.sef24sp4.common.vector.IVector;
 
 public interface IMouseCoordinates extends IVector {
 }
