+++ Common/src/main/java/com/github/sef24sp4/common/vector/BasicVector.java
@@ -1,7 +1,5 @@
 package com.github.sef24sp4.common.vector;
 
-import com.github.sef24sp4.common.interfaces.IVector;
-
 import java.util.Objects;
 
 /**
