+++ Common/src/main/java/module-info.java
@@ -5,4 +5,5 @@ module Common {
 	exports com.github.sef24sp4.common.interfaces;
 	exports com.github.sef24sp4.common.metadata;
 	exports com.github.sef24sp4.common.services;
+	exports com.github.sef24sp4.common.vector;
 }
