+++ Common/src/test/java/com/github/sef24sp4/common/vector/BasicVectorTest.java
@@ -1,6 +1,5 @@
 package com.github.sef24sp4.common.vector;
 
-import com.github.sef24sp4.common.interfaces.IVector;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
