+++ CommonAI/src/main/java/com/github/sef24sp4/common/ai/IPathfindingProvider.java
@@ -1,7 +1,7 @@
 package com.github.sef24sp4.common.ai;
 
 import com.github.sef24sp4.common.entities.IEntity;
-import com.github.sef24sp4.common.interfaces.IVector;
+import com.github.sef24sp4.common.vector.IVector;
 
 public interface IPathfindingProvider {
