+++ Player/src/main/java/com/github/sef24sp4/player/Player.java
@@ -1,6 +1,6 @@
 package com.github.sef24sp4.player;
 
-import com.github.sef24sp4.common.data.Coordinates;
+import com.github.sef24sp4.common.vector.Coordinates;
 import com.github.sef24sp4.common.entities.CommonEntity;
 import com.github.sef24sp4.common.entities.IAttackingEntity;
 import com.github.sef24sp4.common.entities.ICollidableEntity;
