+++ PolygonEnemy/src/main/java/com/github/sef24sp4/polygonenemy/PolygonEnemyPlugin.java
@@ -1,6 +1,6 @@
 package com.github.sef24sp4.polygonenemy;
 
-import com.github.sef24sp4.common.data.Coordinates;
+import com.github.sef24sp4.common.vector.Coordinates;
 import com.github.sef24sp4.common.interfaces.IEntityManager;
 import com.github.sef24sp4.common.interfaces.IGameSettings;
 import com.github.sef24sp4.common.services.IGamePluginService;
